extract metricHttpServer from PrometheusExporter
